refactor: 배너 공통 컴포넌트 변경
