+++ app/lib/constants/banner.ts
@@ -1,1 +1,5 @@
-export const BANNER_IMAGES = ["/당근.svg", "/번개장터.svg", "/중고나라.svg"];
+export const BANNER_IMAGES = [
+  { id: "daangn", url: "/당근.svg" },
+  { id: "bunjang", url: "/번개장터.svg" },
+  { id: "joongna", url: "/중고나라.svg" },
+];
